Restructure project directory
